Separated loggers and applied suggestions
