+++ src/config.toml
@@ -7,19 +7,21 @@ file_server_path = "static" # Change name to front build path
 [server.endpoints]
 pod_data = "/podDataStructure"
 order_data = "/orderStructures"
+uploadable_boards = "/uploadableBoards"
 websocket = "/backend"
 file_server = "/"
 
-[vehicle]
+[vehicle.network]
 tcp_client_tag = "TCP_CLIENT"
 tcp_server_tag = "TCP_SERVER"
 udp_tag = "UDP"
-# network = { mtu = 1500, sniffer_interface = "\\Device\\NPF_{015224CF-DC7F-4F81-AA58-947027697645}" }
-network = { mtu = 1500, sniffer_interface = "\\Device\\NPF_Loopback" }
-[vehicle.messages]
+# sniffer = { mtu = 1500, interface = "lo" }
+sniffer = { mtu = 1500, interface = "\\Device\\NPF_Loopback" }
+# blcu_ack_id = "blcu_ack"
+
+[vehicle.protections]
 fault_id_key = "fault"
 warning_id_key = "warning"
-blcu_ack_id = "blcu_ack"
 
 [excel.download]
 id = "1BEwASubu0el9oQA6PSwVKaNU-Q6gbJ40JR6kgqguKYE"
@@ -33,18 +35,33 @@ board_sheet_prefix = "BOARD "
 table_prefix = "[TABLE] "
 [excel.parse.global]
 address_table = "addresses"
-backend_entry_key = "Backend"
+backend_key = "Backend"
+blcu_address_key = "BLCU"
 units_table = "units"
 ports_table = "ports"
 board_ids_table = "board_ids"
 message_ids_table = "message_ids"
 
-[logger]
-dump_size = 7000
-row_size = 20
-autosave_delay = "10s"
-path = "./log"
+[logger_handler]
 topics = { enable = "logger/enable", state = "logger/state" }
+base_path = "log"
+flush_interval="5s"
+
+[packet_logger]
+file_name = "packets"
+flush_interval="5s"
+
+[value_logger]
+folder_name = "values"
+flush_interval = "5s"
+
+[order_logger]
+file_name = "orders"
+flush_interval="5s"
+
+[protection_logger]
+file_name = "protections"
+flush_interval="5s"
 
 [orders]
 send_topic = "order/send"
