+++ src/value_logger/loggable_value.go
@@ -1,0 +1,33 @@
+package value_logger
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/HyperloopUPV-H8/Backend-H8/packet"
+)
+
+type LoggableValue struct {
+	ValueId   string
+	Value     packet.Value
+	Timestamp time.Time
+}
+
+func (value LoggableValue) Id() string {
+	return value.ValueId
+}
+
+func (value LoggableValue) Log() []string {
+	return []string{
+		value.Timestamp.String(),
+		fmt.Sprintf("%v", value.Value),
+	}
+}
+
+func ToLoggableValue(id string, value packet.Value, timestamp time.Time) LoggableValue {
+	return LoggableValue{
+		ValueId:   id,
+		Value:     value,
+		Timestamp: timestamp,
+	}
+}
